requirements.txt
appdirs==1.4.3
Django==1.11.1
djangorestframework==3.6.3
djangorestframework-jwt==1.10.0
packaging==16.8
PyJWT==1.5.0
pyparsing==2.2.0
pytz==2017.2
requests==2.11.1
six==1.10.0
